chore: move config and global types to project root
