--- src/configs/contact.ts
@@ -1,1 +1,0 @@
-export const CONTACT_EMAIL = "hey@andremarinho.me";
