--- src/configs/site.ts
@@ -1,2 +1,0 @@
-export const SITE_URL = "https://andremarinho.me";
-export const SITE_NAME = "André Marinho";
